+++ sync-for-commerce/RELEASES.md
@@ -549,3 +549,13 @@ Based on:
 - [python v2.1.0] sync-for-commerce
 ### Releases
 - [PyPI v2.1.0] https://pypi.org/project/codat-sync-for-commerce/2.1.0 - sync-for-commerce
+
+## 2023-10-04 00:04:12
+### Changes
+Based on:
+- OpenAPI Doc 1.1 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Sync-Commerce.yaml
+- Speakeasy CLI 1.93.2 (2.144.7) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v2.2.0] sync-for-commerce
+### Releases
+- [PyPI v2.2.0] https://pypi.org/project/codat-sync-for-commerce/2.2.0 - sync-for-commerce